requirements.txt
# VAESTA - Fashion Recommendation System
streamlit>=1.28.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
